Struct fractal_api::v1::types::User

pub struct User { /* fields omitted */ }

Struct that holds all personal information for the user.

Methods

impl User

fn get_id(&self) -> u64

Gets the ID of the user.

fn get_username(&self) -> &str

Gets the username of the user.

fn get_display_name(&self) -> &str

Gets the displayname of the user

fn get_email(&self) -> &str

Gets the email of the user.

fn is_email_confirmed(&self) -> bool

Returns wether the email of the user has been confirmed or not.

fn get_first_name(&self) -> Option<&str>

Gets the first name of the user, if it has been set.

fn is_first_name_confirmed(&self) -> bool

Returns wether the first name of the user has been confirmed or not.

fn get_last_name(&self) -> Option<&str>

Gets the last name of the user, if it has been set.

fn is_last_name_confirmed(&self) -> bool

Returns wether the last name of the user has been confirmed or not.

fn get_device_count(&self) -> u8

Gets the device count of the user.

fn wallet_addresses(&self) -> Iter<WalletAddress>

Gets an iterator through the wallet addresses of the user.

fn get_checking_balance(&self) -> Amount

Gets the checking balance of the user.

fn get_cold_balance(&self) -> Amount

Gets the cold balance of the user.

fn bonds(&self) -> Iter<DateTime<UTC>, u64>

Gets the bonds purchased by the user.

fn get_birthday(&self) -> Option<NaiveDate>

Gets the birthday of the user, if it has been set.

fn is_birthday_confirmed(&self) -> bool

Returns wether the birthday of the user has been confirmed or not.

fn get_phone(&self) -> Option<&str>

Gets the phone of the user, if it has been set.

fn is_phone_confirmed(&self) -> bool

Returns wether the phone of the user has been confirmed or not.

fn get_image_url(&self) -> Option<&str>

Gets the image of the user, if it has been set.

fn get_address(&self) -> Option<&Address>

Gets the address of the user, if it has been set.

fn is_address_confirmed(&self) -> bool

Returns wether the address of the user is confirmed or not.

fn get_sybil_score(&self) -> i8

Gets the sybil score of the user.

fn get_trust_score(&self) -> i8

Gets the trust score of the user.

fn is_enabled(&self) -> bool

Returns wether the user is enabled or not.

fn get_registration_time(&self) -> DateTime<UTC>

Gets the registration time of the user.

fn get_last_activity(&self) -> DateTime<UTC>

Gets the last activity time of the user.

fn is_banned(&self) -> bool

Returns wether the user is banned or not.

fn ban_expiration(&self) -> Option<DateTime<UTC>>

Returns the ban's expiration time, if it has been set.